requirements.txt
Jinja2>=3.0.0
beautifulsoup4>=4.9.0
requests>=2.25.0
pyspellchecker>=0.7.0